chore: Extract config and engine builder modules from turborepo-lib
- Create new `turborepo-config` crate (~1,900 lines) containing:
  - `ConfigurationOptions`, `TurborepoConfigBuilder`, `Error`
  - Environment variable handling (`env.rs`, `override_env.rs`)
  - File-based configuration loading (`file.rs`, `turbo_json.rs`)

- Enhance `turborepo-engine` crate (~1,400 lines added) with:
  - `EngineBuilder<'a, L: TurboJsonLoader>` - generic over loader
  - `BuilderError` and helper error types
  - `TaskDefinitionFromProcessed` trait
  - `TurboJsonLoader` trait abstraction
  - Validation utilities (`validate_task_name`)

- Maintain backward compatibility via re-exports:
  - `turborepo-lib::config::*` re-exports from `turborepo-config`
  - `turborepo-lib::engine::EngineBuilder` re-exports from `turborepo-engine`
  - Type alias `Error = BuilderError` for API compatibility

- Move isolated `test_validate_task_name` test to `turborepo-engine`

This reduces `turborepo-lib` coupling and improves modularity while
ensuring all existing code continues to work through re-exports.
