+++ crates/turborepo-config/src/env.rs
@@ -10,7 +10,7 @@ use turbopath::AbsoluteSystemPathBuf;
 use turborepo_cache::CacheConfig;
 use turborepo_types::{EnvMode, LogOrder, UIMode};
 
-use super::{ConfigurationOptions, Error, ResolvedConfigurationOptions};
+use crate::{ConfigurationOptions, Error, ResolvedConfigurationOptions};
 
 const TURBO_MAPPING: &[(&str, &str)] = [
     ("turbo_api", "api_url"),
@@ -299,7 +299,7 @@ mod test {
     use turborepo_types::LogOrder;
 
     use super::*;
-    use crate::config::{DEFAULT_API_URL, DEFAULT_LOGIN_URL, DEFAULT_TUI_SCROLLBACK_LENGTH};
+    use crate::{DEFAULT_API_URL, DEFAULT_LOGIN_URL, DEFAULT_TUI_SCROLLBACK_LENGTH};
 
     #[test]
     fn test_env_setting() {
