+++ crates/turborepo-lib/Cargo.toml
@@ -128,6 +128,7 @@ turborepo-auth = { path = "../turborepo-auth" }
 turborepo-boundaries = { workspace = true }
 turborepo-cache = { workspace = true }
 turborepo-ci = { workspace = true }
+turborepo-config = { path = "../turborepo-config" }
 turborepo-daemon = { workspace = true }
 turborepo-devtools = { path = "../turborepo-devtools" }
 turborepo-diagnostics = { workspace = true }
