+++ crates/turborepo-lib/src/engine/mod.rs
@@ -1,18 +1,20 @@
 mod builder;
-pub(crate) mod task_inheritance;
 
-pub use builder::{EngineBuilder, Error as BuilderError};
 use miette::{Diagnostic, NamedSource, SourceSpan};
 use thiserror::Error;
 // Building state is used for engine construction
 #[cfg(test)]
 pub use turborepo_engine::Building;
+// Re-export builder types from turborepo-engine
+pub use turborepo_engine::{BuilderError, EngineBuilder};
 // Re-export core types from turborepo-engine
 pub use turborepo_engine::{
     Built, ExecuteError, ExecutionOptions, Message, StopExecution, TaskDefinitionInfo, TaskNode,
 };
 use turborepo_repository::package_graph::{PackageGraph, PackageName};
 use turborepo_types::{TaskDefinition, UIMode};
+// Keep backward compatibility type alias
+pub type Error = BuilderError;
 
 /// Type alias for Engine specialized with TaskDefinition.
 /// This allows existing code to continue using `Engine` without type
