+++ crates/turborepo-lib/src/turbo_json/mod.rs
@@ -41,6 +41,7 @@ pub mod parser {
 
 // Re-export the validator module
 pub mod validator {
+    #[allow(unused_imports)]
     pub use turborepo_turbo_json::validator::*;
 }
 
@@ -48,13 +49,14 @@ pub mod validator {
 use std::collections::HashMap;
 
 pub use loader::{TurboJsonLoader, TurboJsonReader};
-use turbopath::RelativeUnixPath;
+// Re-export TaskDefinitionFromProcessed from turborepo-engine (used by dependent crates)
+#[allow(unused_imports)]
+pub use turborepo_engine::TaskDefinitionFromProcessed;
 use turborepo_errors::Spanned;
 use turborepo_task_id::TaskName;
-use turborepo_types::TaskInputs;
 use turborepo_unescape::UnescapedString;
 
-use crate::{config::Error, run::task_access::TaskAccessTraceFile, task_graph::TaskDefinition};
+use crate::run::task_access::TaskAccessTraceFile;
 
 /// Extension trait for RawTurboJson with turborepo-lib specific functionality
 pub trait RawTurboJsonExt {
@@ -101,121 +103,6 @@ impl RawTurboJsonExt for RawTurboJson {
             tasks: Some(pipeline),
             ..RawTurboJson::default()
         })
-    }
-}
-
-/// Extension trait for creating TaskDefinition from processed task definitions.
-pub trait TaskDefinitionFromProcessed {
-    /// Creates a TaskDefinition from a ProcessedTaskDefinition
-    fn from_processed(
-        processed: ProcessedTaskDefinition,
-        path_to_repo_root: &RelativeUnixPath,
-    ) -> Result<TaskDefinition, Error>;
-
-    /// Helper method for tests that still use RawTaskDefinition
-    #[cfg(test)]
-    fn from_raw(
-        raw_task: RawTaskDefinition,
-        path_to_repo_root: &RelativeUnixPath,
-    ) -> Result<TaskDefinition, Error>;
-}
-
-impl TaskDefinitionFromProcessed for TaskDefinition {
-    fn from_processed(
-        processed: ProcessedTaskDefinition,
-        path_to_repo_root: &RelativeUnixPath,
-    ) -> Result<TaskDefinition, Error> {
-        // Convert outputs with turbo_root resolution
-        let outputs = processed
-            .outputs
-            .map(|outputs| task_outputs_from_processed(outputs, path_to_repo_root))
-            .transpose()?
-            .unwrap_or_default();
-
-        let cache = processed.cache.is_none_or(|c| c.into_inner());
-        let interactive = processed
-            .interactive
-            .as_ref()
-            .map(|value| value.value)
-            .unwrap_or_default();
-
-        if let Some(interactive) = &processed.interactive {
-            let (span, text) = interactive.span_and_text("turbo.json");
-            if cache && interactive.value {
-                return Err(Error::TurboJsonError(
-                    turborepo_turbo_json::Error::InteractiveNoCacheable { span, text },
-                ));
-            }
-        }
-
-        let persistent = *processed.persistent.unwrap_or_default();
-        let interruptible = processed.interruptible.unwrap_or_default();
-        if *interruptible && !persistent {
-            let (span, text) = interruptible.span_and_text("turbo.json");
-            return Err(Error::TurboJsonError(
-                turborepo_turbo_json::Error::InterruptibleButNotPersistent { span, text },
-            ));
-        }
-
-        let mut topological_dependencies: Vec<Spanned<TaskName>> = Vec::new();
-        let mut task_dependencies: Vec<Spanned<TaskName>> = Vec::new();
-        if let Some(depends_on) = processed.depends_on {
-            for dependency in depends_on.deps {
-                let (dependency, depspan) = dependency.split();
-                let dependency: String = dependency.into();
-                if let Some(topo_dependency) =
-                    dependency.strip_prefix(TOPOLOGICAL_PIPELINE_DELIMITER)
-                {
-                    topological_dependencies.push(depspan.to(topo_dependency.to_string().into()));
-                } else {
-                    task_dependencies.push(depspan.to(dependency.into()));
-                }
-            }
-        }
-
-        task_dependencies.sort_by(|a, b| a.value.cmp(&b.value));
-        topological_dependencies.sort_by(|a, b| a.value.cmp(&b.value));
-
-        let env = processed.env.map(|env| env.vars).unwrap_or_default();
-
-        // Convert inputs with turbo_root resolution
-        let inputs = processed
-            .inputs
-            .map(|inputs| TaskInputs::from_processed(inputs, path_to_repo_root))
-            .unwrap_or_default();
-
-        let pass_through_env = processed.pass_through_env.map(|env| env.vars);
-
-        let with = processed.with.map(|with_tasks| with_tasks.tasks);
-
-        Ok(TaskDefinition {
-            outputs,
-            cache,
-            topological_dependencies,
-            task_dependencies,
-            env,
-            inputs,
-            pass_through_env,
-            output_logs: *processed.output_logs.unwrap_or_default(),
-            persistent,
-            interruptible: *interruptible,
-            interactive,
-            env_mode: processed.env_mode.map(|mode| *mode.as_inner()),
-            with,
-        })
-    }
-
-    #[cfg(test)]
-    fn from_raw(
-        raw_task: RawTaskDefinition,
-        path_to_repo_root: &RelativeUnixPath,
-    ) -> Result<TaskDefinition, Error> {
-        // Use default FutureFlags for backward compatibility
-        let processed = ProcessedTaskDefinition::from_raw(raw_task, &FutureFlags::default())?;
-        <TaskDefinition as TaskDefinitionFromProcessed>::from_processed(
-            processed,
-            path_to_repo_root,
-        )
     }
 }
